Finish rework of ProjectTimline (ProjectWebParts)
